WIP: Make DGM and fuzz tester work together
