+++ p2panda-core/src/hash.rs
@@ -167,13 +167,10 @@ mod tests {
     fn hashing() {
         let hash = Hash::new([1, 2, 3]);
 
-        assert_eq!(
-            hash.as_bytes(),
-            &[
-                177, 119, 236, 27, 242, 109, 251, 59, 112, 16, 212, 115, 230, 212, 71, 19, 178,
-                155, 118, 91, 153, 198, 230, 14, 203, 250, 231, 66, 222, 73, 101, 67
-            ]
-        );
+        assert_eq!(hash.as_bytes(), &[
+            177, 119, 236, 27, 242, 109, 251, 59, 112, 16, 212, 115, 230, 212, 71, 19, 178, 155,
+            118, 91, 153, 198, 230, 14, 203, 250, 231, 66, 222, 73, 101, 67
+        ]);
     }
 
     #[test]
