+++ p2panda-group/src/message_scheme/group.rs
@@ -2,6 +2,7 @@
 
 #![allow(clippy::type_complexity)]
 use std::collections::{HashMap, HashSet, VecDeque};
+use std::hash::Hash as StdHash;
 use std::marker::PhantomData;
 
 use serde::{Deserialize, Serialize};
@@ -131,6 +132,10 @@ where
             return Err(GroupError::NotAddOurselves);
         }
 
+        if Self::members(&y)?.contains(&added) {
+            return Err(GroupError::AddedExistsAlready(added));
+        }
+
         // Add a new member to the group.
         let (y_dcgka_i, pre) = Dcgka::add(y.dcgka, added, rng)?;
         y.dcgka = y_dcgka_i;
@@ -146,6 +151,10 @@ where
     ) -> GroupResult<ORD::Message, ID, OP, PKI, DGM, KMG, ORD> {
         if y.ratchet.is_none() {
             return Err(GroupError::GroupNotYetEstablished);
+        }
+
+        if !Self::members(&y)?.contains(&removed) {
+            return Err(GroupError::InexistentRemovedMember(removed));
         }
 
         // Remove a member from the group.
@@ -182,7 +191,11 @@ where
         mut y: GroupState<ID, OP, PKI, DGM, KMG, ORD>,
         message: &ORD::Message,
         rng: &Rng,
-    ) -> GroupResult<Vec<GroupOutput<ID, OP, DGM, ORD>>, ID, OP, PKI, DGM, KMG, ORD> {
+    ) -> GroupResult<Option<GroupOutput<ID, OP, DGM, ORD>>, ID, OP, PKI, DGM, KMG, ORD> {
+        // Remember current members state so we can calculate a diff after processing these
+        // messages.
+        let members_pre = Self::members(&y)?;
+
         let message_type = message.message_type();
         let is_established = y.ratchet.is_some();
         let mut is_create_or_welcome = false;
@@ -216,7 +229,7 @@ where
             // We're receiving control- and application messages for this group but we haven't
             // joined yet. We keep these messages for later. We don't know yet when we will join
             // the group and which of these messages we can process afterwards.
-            return Ok((y, vec![]));
+            return Ok((y, None));
         }
 
         if !is_established && is_create_or_welcome {
@@ -231,10 +244,15 @@ where
             // Always process welcome message first before anything else.
             let (y_i, result) = Self::process_ready(y, &message, rng)?;
 
-            return Ok((y_i, result.map_or(vec![], |output| vec![output])));
+            let members_post = Self::members(&y_i)?;
+
+            return Ok((
+                y_i,
+                result.map(|output| GroupOutput::new(vec![output], members_pre, members_post)),
+            ));
         }
 
-        let mut results = Vec::new();
+        let mut events = Vec::new();
         let mut y_loop = y;
 
         let mut control_messages = VecDeque::new();
@@ -274,7 +292,7 @@ where
             let (y_next, result) = Self::process_ready(y_loop, &message, rng)?;
             y_loop = y_next;
             if let Some(message) = result {
-                results.push(message);
+                events.push(message);
             }
         }
 
@@ -283,11 +301,16 @@ where
             let (y_next, result) = Self::process_ready(y_loop, &message, rng)?;
             y_loop = y_next;
             if let Some(message) = result {
-                results.push(message);
+                events.push(message);
             }
         }
 
-        Ok((y_loop, results))
+        let members_post = Self::members(&y_loop)?;
+
+        Ok((
+            y_loop,
+            Some(GroupOutput::new(events, members_pre, members_post)),
+        ))
     }
 
     /// Encrypts application message towards the current group.
@@ -355,7 +378,7 @@ where
         y: GroupState<ID, OP, PKI, DGM, KMG, ORD>,
         message: &ORD::Message,
         rng: &Rng,
-    ) -> GroupResult<Option<GroupOutput<ID, OP, DGM, ORD>>, ID, OP, PKI, DGM, KMG, ORD> {
+    ) -> GroupResult<Option<GroupEvent<ID, OP, DGM, ORD>>, ID, OP, PKI, DGM, KMG, ORD> {
         match message.message_type() {
             ForwardSecureMessageType::Control(control_message) => {
                 let direct_message = message
@@ -375,9 +398,9 @@ where
                 // Check if processing this message removed us from the group.
                 let is_removed = !Self::members(&y_i)?.contains(&y_i.my_id);
                 if is_removed {
-                    Ok((y_i, Some(GroupOutput::Removed)))
+                    Ok((y_i, Some(GroupEvent::RemovedOurselves)))
                 } else {
-                    Ok((y_i, output.map(|msg| GroupOutput::Control(msg))))
+                    Ok((y_i, output.map(|msg| GroupEvent::Control(msg))))
                 }
             }
             ForwardSecureMessageType::Application {
@@ -385,7 +408,7 @@ where
                 generation,
             } => {
                 let (y_i, plaintext) = Self::decrypt(y, message.sender(), ciphertext, generation)?;
-                Ok((y_i, Some(GroupOutput::Application { plaintext })))
+                Ok((y_i, Some(GroupEvent::Application { plaintext })))
             }
         }
     }
@@ -484,8 +507,50 @@ where
 pub type GroupResult<T, ID, OP, PKI, DGM, KMG, ORD> =
     Result<(GroupState<ID, OP, PKI, DGM, KMG, ORD>, T), GroupError<ID, OP, PKI, DGM, KMG, ORD>>;
 
+#[derive(Clone, Default)]
+pub struct GroupOutput<ID, OP, DGM, ORD>
+where
+    ID: Clone + Eq + PartialEq + StdHash,
+    DGM: AckedGroupMembership<ID, OP>,
+    ORD: ForwardSecureOrdering<ID, OP, DGM>,
+{
+    pub events: Vec<GroupEvent<ID, OP, DGM, ORD>>,
+    pub removed_members: HashSet<ID>,
+    pub added_members: HashSet<ID>,
+}
+
+impl<ID, OP, DGM, ORD> GroupOutput<ID, OP, DGM, ORD>
+where
+    ID: Clone + Eq + PartialEq + StdHash,
+    DGM: AckedGroupMembership<ID, OP>,
+    ORD: ForwardSecureOrdering<ID, OP, DGM>,
+{
+    pub(crate) fn new(
+        events: Vec<GroupEvent<ID, OP, DGM, ORD>>,
+        pre_members: HashSet<ID>,
+        post_members: HashSet<ID>,
+    ) -> Self {
+        let added_members = post_members.difference(&pre_members).cloned().collect();
+        let removed_members = pre_members.difference(&post_members).cloned().collect();
+
+        GroupOutput {
+            events,
+            removed_members,
+            added_members,
+        }
+    }
+
+    pub(crate) fn from_events(events: Vec<GroupEvent<ID, OP, DGM, ORD>>) -> Self {
+        GroupOutput {
+            events,
+            removed_members: HashSet::new(),
+            added_members: HashSet::new(),
+        }
+    }
+}
+
 #[derive(Clone, Debug, PartialEq, Eq)]
-pub enum GroupOutput<ID, OP, DGM, ORD>
+pub enum GroupEvent<ID, OP, DGM, ORD>
 where
     DGM: AckedGroupMembership<ID, OP>,
     ORD: ForwardSecureOrdering<ID, OP, DGM>,
@@ -497,7 +562,7 @@ where
     Application { plaintext: Vec<u8> },
 
     /// Signal that we've been removed from the group.
-    Removed,
+    RemovedOurselves,
 }
 
 #[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
@@ -565,6 +630,12 @@ where
 
     #[error("decryption ratchet not established yet to process the message from {0} @{1}")]
     DecryptionRachetUnavailable(ID, Generation),
+
+    #[error("to-be-added member {0} is already part of the group")]
+    AddedExistsAlready(ID),
+
+    #[error("to-be-removed member {0} is not part of the group")]
+    InexistentRemovedMember(ID),
 }
 
 #[cfg(test)]
