+++ p2panda-group/src/message_scheme/test_utils/dcgka.rs
@@ -3,7 +3,7 @@
 use std::collections::{HashMap, HashSet};
 
 use crate::crypto::x25519::SecretKey;
-use crate::message_scheme::acked_dgm::test_utils::AckedTestDgm;
+use crate::message_scheme::test_utils::dgm::AckedTestDgm;
 use crate::message_scheme::{
     ControlMessage, Dcgka, DcgkaState, DirectMessage, DirectMessageType, OperationOutput,
     ProcessOutput, UpdateSecret,
